requirements.txt
# Development
ruff>=0.2.1             # Linting
black>=24.1.1           # Code formatting
mypy>=1.8.0             # Type checking
pydantic-settings>=2.0.0

# Data Processing
pandas>=2.0.0
scikit-learn>=1.0.0
matplotlib>=3.7.0
seaborn>=0.13.0
optuna>=3.0.0

pytest>=7.0.0

# Add these to your existing requirements
pyyaml>=6.0.1